requirements.txt
line-bot-sdk==1.12.1
openai
boto3
